+++ tests/test_jump2d.lua
@@ -585,7 +585,7 @@ T['start()']['handles very big `view.n_steps_ahead`'] = function()
 end
 
 T['start()']['handles overlapping multi-step labels'] = function()
-  child.lua([[MiniJump2d.config.spotter = MiniJump2d.gen_pattern_spotter('.')]])
+  child.lua([[MiniJump2d.config.spotter = MiniJump2d.gen_spotter.pattern('.')]])
   start({ labels = 'jk', view = { n_steps_ahead = 2 } })
   child.expect_screenshot()
 
@@ -901,7 +901,9 @@ T['stop()']['works even if not jumping'] = function()
   eq(get_cursor(), init_cursor)
 end
 
-T['gen_pattern_spotter()'] = new_set({
+T['gen_spotter'] = new_set()
+
+T['gen_spotter']['pattern()'] = new_set({
   hooks = {
     pre_case = function()
       set_lines({ 'xxx x_x x.x xxx' })
@@ -912,25 +914,25 @@ T['gen_pattern_spotter()'] = new_set({
 
 local start_gen_pattern = function(pattern, side)
   local command = string.format(
-    [[MiniJump2d.start({ spotter = MiniJump2d.gen_pattern_spotter(%s, %s) })]],
+    [[MiniJump2d.start({ spotter = MiniJump2d.gen_spotter.pattern(%s, %s) })]],
     vim.inspect(pattern),
     vim.inspect(side)
   )
   child.lua(command)
 end
 
-T['gen_pattern_spotter()']['works'] = function()
+T['gen_spotter']['pattern()']['works'] = function()
   start_gen_pattern(nil, nil)
   -- By default it matches group of non-whitespace non-punctuation
   child.expect_screenshot()
 end
 
-T['gen_pattern_spotter()']['respects `pattern` argument'] = function()
+T['gen_spotter']['pattern()']['respects `pattern` argument'] = function()
   start_gen_pattern('%s', nil)
   child.expect_screenshot()
 end
 
-T['gen_pattern_spotter()']['respects `side` argument'] = new_set({
+T['gen_spotter']['pattern()']['respects `side` argument'] = new_set({
   parametrize = { { '%S+', 'start' }, { '%S+', 'end' }, { '.().', 'none' } },
 }, {
   test = function(pattern, side)
@@ -939,7 +941,7 @@ T['gen_pattern_spotter()']['respects `side` argument'] = new_set({
   end,
 })
 
-T['gen_pattern_spotter()']['handles patterns with "^" and "$"'] = new_set({
+T['gen_spotter']['pattern()']['handles patterns with "^" and "$"'] = new_set({
   parametrize = {
     { '^...', 'start' },
     { '^...', 'end' },
@@ -956,7 +958,7 @@ T['gen_pattern_spotter()']['handles patterns with "^" and "$"'] = new_set({
   end,
 })
 
-T['gen_pattern_spotter()']['works with multibyte characters'] = function()
+T['gen_spotter']['pattern()']['works with multibyte characters'] = function()
   set_lines({ 'ы ыыы ы_ы ыы' })
   start_gen_pattern('%S')
   child.expect_screenshot()
@@ -969,31 +971,31 @@ T['gen_pattern_spotter()']['works with multibyte characters'] = function()
   child.expect_screenshot()
 end
 
-T['gen_pattern_spotter()']['works in edge cases'] = function()
+T['gen_spotter']['pattern()']['works in edge cases'] = function()
   start_gen_pattern('.%f[%W]')
   child.expect_screenshot()
 end
 
-T['gen_union_spotter()'] = new_set()
+T['gen_spotter']['union()'] = new_set()
 
-T['gen_union_spotter()']['works'] = function()
+T['gen_spotter']['union()']['works'] = function()
   child.set_size(5, 25)
 
   child.lua([[
-    local nonblank_start = MiniJump2d.gen_pattern_spotter('%S+', 'start')
+    local nonblank_start = MiniJump2d.gen_spotter.pattern('%S+', 'start')
     _G.args_log = {}
     _G.spotter_1 = function(...)
       table.insert(_G.args_log, { ... })
       return nonblank_start(...)
     end
 
-    local word_start = MiniJump2d.gen_pattern_spotter('%w+', 'start')
+    local word_start = MiniJump2d.gen_spotter.pattern('%w+', 'start')
     _G.spotter_2 = function(...)
       table.insert(_G.args_log, { ... })
       return word_start(...)
     end
 
-    _G.union_spotter = MiniJump2d.gen_union_spotter(_G.spotter_1, _G.spotter_2)
+    _G.union_spotter = MiniJump2d.gen_spotter.union(_G.spotter_1, _G.spotter_2)
   ]])
 
   set_lines({ 'xxx x_x x_x xxx' })
@@ -1001,15 +1003,15 @@ T['gen_union_spotter()']['works'] = function()
   child.expect_screenshot()
 end
 
-T['gen_union_spotter()']['validates arguments'] = function()
+T['gen_spotter']['union()']['validates arguments'] = function()
   expect.error(
-    function() child.lua('MiniJump2d.gen_union_spotter(function() end, 1, function() end)') end,
+    function() child.lua('MiniJump2d.gen_spotter.union(function() end, 1, function() end)') end,
     'All.*callable'
   )
 end
 
-T['gen_union_spotter()']['works with no arguments'] = function()
-  child.lua('_G.spotter = MiniJump2d.gen_union_spotter()')
+T['gen_spotter']['union()']['works with no arguments'] = function()
+  child.lua('_G.spotter = MiniJump2d.gen_spotter.union()')
 
   set_lines({ 'xxx x_x x_x xxx' })
   eq(child.lua_get('_G.spotter(1, {})'), {})