+++ CHANGELOG.md
@@ -7,6 +7,16 @@ There are following change types:
 
 # Version 0.16.0.9000
 
+## mini.jump2d
+
+### Refine
+
+- Move `gen_xxx_spotter` into separate `gen_spotter` table for consistency with other modules:
+    - `gen_pattern_spotter` -> `gen_spotter.pattern`
+    - `gen_union_spotter` -> `gen_spotter.union`
+
+    The `gen_xxx_spotter` functions will work (with warning) until at least next release.
+
 ## mini.pick
 
 ### Expand